add 404 page
create a component for when no pages match, then import Switch from react-router-dom and put your routes inside of opening and closing <Switch> tags so it will only render the first route that matches. You would then put the Route for when there is no match (your 404 / error page) as the last route.